SYNC OFFLINE GHANTA GADI - IF SOMETHING WENT WRONG ISSUE TRIGGERS ( DATA NOT DELETING )

+++ swachbharatabhiyanlibrary/src/main/java/com/appynitty/swachbharatabhiyanlibrary/adapters/connection/SyncOfflineAdapterClass.java
@@ -85,13 +85,11 @@ public class SyncOfflineAdapterClass {
 
                                 } else {
                                     Log.d(TAG, "onResponse: " + response);
-                                    // AUtils.warning(mContext, response.message(), Toast.LENGTH_SHORT);
+                                    AUtils.warning(mContext, response.message(), Toast.LENGTH_SHORT);
                                     Log.i("SyncOfflineClass", "onFailureCallback: Response Code-" + response.code());
                                     Log.i("SyncOfflineClass", "onFailureCallback: Response Code-" + response.message());
-//                                    AUtils.isSyncOfflineDataRequestEnable = false;
-//                                    Prefs.putBoolean(AUtils.isSyncingOn, false);
-
-                                    SyncOfflineData();
+                                    AUtils.isSyncOfflineDataRequestEnable = false;
+                                    //            Prefs.putBoolean(AUtils.isSyncingOn, false);
                                     syncOfflineListener.onFailureCallback();
                                 }
                             }
@@ -100,17 +98,16 @@ public class SyncOfflineAdapterClass {
                             public void onFailure(Call<List<OfflineGcResultPojo>> call, Throwable t) {
                                 Log.i(AUtils.TAG_HTTP_RESPONSE, "onFailureCallback: Response Code-" + t.getMessage());
                                 AUtils.warning(mContext, t.getMessage(), Toast.LENGTH_SHORT);
-                                //   AUtils.isSyncOfflineDataRequestEnable = false;
-                                // Prefs.putBoolean(AUtils.isSyncingOn, false);
+                                AUtils.isSyncOfflineDataRequestEnable = false;
+                                Prefs.putBoolean(AUtils.isSyncingOn, false);
                                 syncOfflineListener.onErrorCallback();
-                                SyncOfflineData();
                                 Log.i("RESPONSE_CODE", "onResponse: " + t.getMessage());
 
                             }
 
                         });
             } else {
-                Prefs.putBoolean(AUtils.isSyncingOn, false);
+                //       Prefs.putBoolean(AUtils.isSyncingOn, false);
                 if (syncOfflineListener != null) {
                     syncOfflineListener.onSuccessCallback();
 
@@ -151,8 +148,8 @@ public class SyncOfflineAdapterClass {
                                 areaWarningListener.onError(result.getMessage());
 
                         } else {
-                            syncOfflineList.clear();
-                            syncOfflineRepository.deleteCompleteSyncTableData();
+                            //   syncOfflineList.clear();
+                            //    syncOfflineRepository.deleteCompleteSyncTableData();
                             if (Prefs.getString(AUtils.LANGUAGE_NAME, AUtils.DEFAULT_LANGUAGE_ID).equals(AUtils.LanguageConstants.MARATHI))
                                 AUtils.error(mContext, result.getMessageMar());
                             else
@@ -162,17 +159,29 @@ public class SyncOfflineAdapterClass {
                     }
 
                     if (Integer.parseInt(result.getID()) != 0) {
-                        int deleteCount = syncOfflineRepository.deleteSyncTableData(result.getID());
-                        if (deleteCount == 0) {
-                            offset = String.valueOf(Integer.parseInt(offset) + 1);
+
+//                        if (result.getMessage().contains("scanned")) {
+//
+//                        } else if (result.getMessage().contains("Invalid")) {
+//
+//                        } else
+
+                        //CHANGES BY SANATH GOSAVI
+                        if (!result.getMessage().contains("wrong")) {
+                            int deleteCount = syncOfflineRepository.deleteSyncTableData(result.getID());
+                            if (deleteCount == 0) {
+                                offset = String.valueOf(Integer.parseInt(offset) + 1);
+                            }
+                            for (int i = 0; i < syncOfflineList.size(); i++) {
+                                if (syncOfflineList.get(i).getOfflineID().equals(result.getID())) {
+                                    syncOfflineList.remove(i);
+                                    break;
+                                }
+                            }
                         }
+
                     }
-                    for (int i = 0; i < syncOfflineList.size(); i++) {
-                        if (syncOfflineList.get(i).getOfflineID().equals(result.getID())) {
-                            syncOfflineList.remove(i);
-                            break;
-                        }
-                    }
+
                 } else
                     offset = String.valueOf(Integer.parseInt(offset) + 1);
             }
@@ -186,17 +195,14 @@ public class SyncOfflineAdapterClass {
                 }
             }
             if (isSuccess) {
-                if (syncOfflineRepository.fetchCollectionCount().size() == 0) {
+                if (syncOfflineRepository.fetchCollectionCount().size() == 0)
                     AUtils.success(mContext, mContext.getResources().getString(R.string.success_offline_sync));
-                    AUtils.isSyncOfflineDataRequestEnable = false;
-                    Prefs.putBoolean(AUtils.isSyncingOn, false);
-                }
-
             }
 
         }
 
-
+        AUtils.isSyncOfflineDataRequestEnable = false;
+        Prefs.putBoolean(AUtils.isSyncingOn, false);
     }
 
     private void setOfflineData() {
